Move partition list
