+++ perun/utils/__init__.py
@@ -410,25 +410,3 @@ def merge_dictionaries(*args: dict[Any, Any]) -> dict[Any, Any]:
     for dictionary in args:
         res.update(dictionary)
     return res
-
-
-def partition_list(
-    input_list: Iterable[Any], condition: Callable[[Any], bool]
-) -> tuple[list[Any], list[Any]]:
-    """Utility function for list partitioning on a condition so that the list is not iterated
-    twice and the condition is evaluated only once.
-
-    Based on a SO answer featuring multiple methods and their performance comparison:
-    'https://stackoverflow.com/a/31448772'
-
-    :param iterator input_list: the input list to be partitioned
-    :param function condition: the condition that should be evaluated on every list item
-    :return tuple: (list of items evaluated to True, list of items evaluated to False)
-    """
-    good, bad = [], []
-    for item in input_list:
-        if condition(item):
-            good.append(item)
-        else:
-            bad.append(item)
-    return good, bad
